+++ pkg/PHANE-file-formats/src/reader/File/Format/TransitionCostMatrix/Reader.hs
@@ -1,0 +1,11 @@
+{- |
+Functions for for parsing TCM files into an alphabet and native matrix form.
+-}
+module File.Format.TransitionCostMatrix.Reader (
+    FileFormatTCM (..),
+    tcmStreamReader,
+) where
+
+import File.Format.TransitionCostMatrix.Parser
+import File.Format.TransitionCostMatrix.Types
+