+++ pkg/PHANE-measure-class/src/Measure/Range.hs
@@ -1,8 +1,14 @@
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE Safe #-}
 
+{- |
+The range of a measure which consists of an upper and lower bound.
+-}
 module Measure.Range (
+    -- * Type-class
     MeasurableRange (..),
+
+    -- * Associated type-alias
     InclusiveRange,
 ) where
 
@@ -10,7 +16,7 @@ import Data.Ix
 
 
 {- |
-The inclusive lower and upper bounds of a range, respectively.
+The /inclusive/ lower and upper bounds of a range, respectively.
 -}
 type InclusiveRange i = (i, i)